Add negative unit test to prevent insurances from being removed if they're being used by customers

+++ Insurance/Insurance.WebAPI/Services/InsuranceService.cs
@@ -1,6 +1,7 @@
 ﻿using Insurance.Core.Exceptions;
 using Insurance.Core.Models;
 using Insurance.Repository;
+using Microsoft.EntityFrameworkCore.Internal;
 using System;
 using System.Collections.Generic;
 
@@ -50,6 +51,11 @@ namespace Insurance.WebAPI.Services
 
         public void Delete(int insuranceId)
         {
+            var customersByInsurance = _insuranceRepository.GetCustomersByInsurance(insuranceId);
+
+            if (customersByInsurance.Any())
+                throw new InsuranceIsBeingUsedByCustomersException(insuranceId, customersByInsurance);
+
             _insuranceRepository.Delete(insuranceId);
         }
